disable fatal warnings where we undef an object in tests
When you die in a DESTROY, it is normally converted to a warning.  Fatal
warnings can promote that to an error again, which causes a variety of
problems.  On some versions of perl this fatalization in based on the
warning hints of the code throwing the error.  On others, it is based on
the location of the code triggering the DESTROY.  Make sure both cases
are covered in the tests.